Rename initiative breadcrumbs to avoid confusing with new breadcrumbs
